feat(404): added correct dicord link on 404 page. Have also renamed file with typo.
